requirements.txt
asgiref==3.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.0.2
django-allauth==0.51.0
django-cache-url==3.4.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rosetta==0.9.8
environs==9.5.0
idna==3.4
jdatetime==4.1.0
marshmallow==3.18.0
oauthlib==3.2.1
packaging==21.3
Pillow==9.3.0
polib==1.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2021.5
urllib3==1.26.12
